Start refactoring component's distribution to improve UX and DX
This refactoring improves two pain points:
 - Finding the correct readme file for the components. For example, some user would previously go to the tree-view-item folder to see what the item's properties are to understand its usage. Since the Tree View has only one main component (the ch-tree-view-render), users only need to view the ch-tree-view-render's readme file.

 - As Chameleon grows, the number of components will increase and traveling back and forth to the "renders" folder is a pain point. It's better to have only one folder with the implementation of the control.

So, starting from this PR, components that have an asossiated render will one be placed in one folder with the following structure (example):
  components
    my-custom-element/
      internal/
        my-custom-element-item/
          my-custom-element-item.tsx
          my-custom-element-item.scss
          README.md
        my-custom-element-view/
          my-custom-element-view.tsx
          my-custom-element-view.scss
          README.md
        README.md (explains that all files in this folder are internal and the only readme to be see is in the previous folder)
      my-custom-element-render.tsx
      my-custom-element-render.scss
      README.md
