+++ src/components/tree-view/internal/tree-view-item/tree-view-item.scss
@@ -1,4 +1,4 @@
-@import "../../../common/base";
+@import "../../../../common/base";
 
 @include box-sizing();
