+++ src/components/tree-view/internal/tree-view/types.ts
@@ -1,4 +1,4 @@
-import { GxDataTransferInfo } from "../../../common/types";
+import { GxDataTransferInfo } from "../../../../common/types";
 
 export type CheckedTreeItemInfo = {
   id: string;
